change data type of item amount from BIGINT to FLOAT

+++ db/migrations/20230116151613-create-expense.js
@@ -13,7 +13,7 @@ module.exports = {
         type: Sequelize.STRING,
       },
       amount: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.FLOAT,
       },
       payer_id: {
         type: Sequelize.INTEGER,
